requirements.txt
# Core dependencies
crewai>=0.193.0
crewai-tools>=0.12.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Streamlit
streamlit>=1.31.0

# LLM
openai>=1.12.0
langchain-openai>=0.3.0

# Utilities
pydantic>=2.6.0

# Note: Compatible with Python 3.10 - 3.13
# Tested with Python 3.13.3

